[Close #224] Show action below PostsController
Move `Newpoc::Action::Post::Show` to `PostsController::Action::Show`,
with some factoring out of common code.

RSpec: 704 examples, 0 failures
RuboCop: 239 files inspected, no offenses detected
