--- lib/newpoc-action-post-show/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
