--- lib/newpoc-action-post-show/lib/newpoc/action/post/show/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module Post
-      # Retrieves a Post on behalf of current user, for `new_poc`.
-      class Show
-        VERSION = '0.0.1'
-      end
-    end
-  end
-end
